Separate publisher expectation tests over several files
